OverviewUtil in progress

+++ app/src/main/java/cz/cuni/mff/nutritionalassistant/foodtypes/Product.java
@@ -13,12 +13,12 @@ public class Product extends Food {
     private String brandName;
     // TODO look up final keyword, because I don't have to change internals of these Lists
     // zero indexed elements are reference values (NH are bound to them)
-    private List<Integer> servingQuantity;
+    private List<Float> servingQuantity;
     private List<String> servingUnit;
-    private List<Integer> servingWeight;
+    private List<Float> servingWeight;
 
     Product(String foodName, String thumbnailURL, float calories, float fats, float carbohydrates, float proteins, FoodType foodType,
-            String brandName, List<Integer> servingQuantity, List<String> servingUnit, List<Integer> servingWeight) {
+            String brandName, List<Float> servingQuantity, List<String> servingUnit, List<Float> servingWeight) {
 
         super(foodName, thumbnailURL, calories, fats, carbohydrates, proteins, foodType);
         this.brandName = brandName;
